colibri-vf: switch to zImage inside root fs
Due to U-Boot udpate we need to create one big UBI volume. Since
attaching the UBI volume is the expensive operation (compared to
mount the UBIFS inside that volume), it is relatively cheep to
load the kernel directly from the root filesystems /boot directory
instead of creating a own volume for the kernel. This also makes
it simpler to load a mainline kernel (along with the required dtb
file).
However, we loose the integrity check of the uImage. But since
there were no measures in place to handle a invalid image, we do
not loose functionality here. In case we need integrity check, we
should consider switching to the new FIT image format...