base-files/u-boot-toradex-fw-utils: move MMC unlock
The mmc unlock function causes troubles when using Poky reference
distribution since Poky uses the busybox shell:
/etc/rc5.d/S09xserver-nodm: /etc/profile.d/fw_unlock_mmc.sh: line 2: syntax error: unexpected "("

This avoids X getting at all.

This patch does not solve the underlying problem, but works around it
by moving the script to the u-boot-toradex-fw-utils recipe, which is
not installed by the Poky images by default.