u-boot-fw-utils: correct blocks reserved for environment
While we use only one erase block for the U-Boot environment, we
reserve four in case there are bad blocks in the environment area.
Specify the number of blocks so that the Linux utilities are able
to read/write the environment even if the first block of the
environment is bad.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>
Acked-by: Max Krummenacher <max.krummenacher@toradex.com>