linux-toradex-rt_4.9-1.0.x: add kernel recipe with PREEMPT RT
Add kernel recipe which applies the PREEMPT RT patchset. Patches
used in the 4.1 kernel are either obsolete or have been updated.
The patch "Work around CPU stalls in the imx-sdma driver" seems
not to be necessary. The kernel did not show the reported stack
trace after one night of cyclictest while audio running.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>
Acked-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>