imx-loader: fix imx_usb configuration files
The string "header" refers to which imx header the utility should
try to use. Some binaries seem to have 2 headers, and header2
refers to the second.

Our binaries have only 1 header, but imx_usb seems to fall back in
that case. This might not be the case with newer imx_usb utilities.

Specifying just "header" is correct and works in all version of
imx_usb.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>
Acked-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>