linux-toradex-rt: make sure correct defconfig is chosen
The linux-toradex and linux-toradex-rt recipe adds a directory
with valid defconfigs to FILESEXTRAPATHS. Bitbake chooses the
first defconfig which it founds, which in that case is the one
from linux-toradex...

Include linux-toradex first, and then prepend the path from the
linux-toradex-rt recipe to make sure the defconfig for the -rt
kernel will be used.

However, in this scenario the PV variable can not be used since
linux-toradex_4.9-1.0.x.bb appends PV with image version and git
hash. Avoid using variables to avoid any ambiguity.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>
Acked-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>