imx-gpu-viv: drop the -dev dependency on libgl-mesa-dev
When building the SDK one gets a warning and most -dev packages do not
get installed.

| * Problem 1/1:
| *   - package libgl-imx-dev-1:6.2.2.p0-aarch32-r0.armv7at2hf-neon-mx6q
dl requires libgl-mesa-dev, but none of the providers can be installed

The opkg file for libgl-mesa-dev exist but somehow is not considered for
installation. Not depending on it at least installs the other -dev
packages.

See also:
https://www.toradex.com/community/questions/19849/bitbake-do-rootfs-failed-with-rocko-and-dev-pkgs-i.html

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>
Acked-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>