machines: i.mx 8/8X/8MM: keep tune pinned to aarch64
meta-freescale started to set specialised tunes for the i.MX8 based
machines. Override that in our machine specific files so that the
user space packages all have the same ARCH 'aarch64' and need not be
built for each CPU flavour individually.
However change the tune include file so that one could change to one of
the more speciallized tune if one wants to do so at the expense of
additional compile time.

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>