machines: u-boot: use full config name
U-Boot used to be configured by 'make <board>_config', however with the
move integration of the Kconfig system that changed to use a config file
ending in _defconfig. Usually those names are <board>_defconfig.
So nowadays one configures with 'make <whatever>_defconfig'.
If make is called with <whatever>_config that is silently changed to
<whatever>_defconfig.
Change our config targets to the new name to prevent confusion.

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>