imx-sc-firmware: do not use autorev for zeus anymore
The release pinned at git-hash ca3cbc72 is the last one for BSP 4.0. We
want to go on with SCFW to version 1.5.0 which has a new API and is
not recommended for kernel and u-boot used on zeus branch.

Drop autorev so we can merge new stuff for dunfell into scfw.

Related-to: ELB-2798
Signed-off-by: Philippe Schenker <philippe.schenker@toradex.com>