weston: revert use systemd notify
Commit 4efdcc1090 ("weston: Use systemd notify,") in openembedded-core
introduced systemd notify for weston.

This seems, at least in our configuration, not to work, both with openembedded
weston 8.0.0 and NXP's fork weston 9.0.0.imx.
Weston starts but systemd does kill it after a timeout as it does
not receive a notification from the spawned deamon.

Comment the relevant lines in the service file, so that systemd
does not expect any notification.

Compare also with https://github.com/Freescale/meta-freescale/issues/901

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>