alsa-state: change asound.conf files to use 48kHz for verdin-imx8 boards
Configure alsa audio to use dmix/dsnoop at 48 kHz.

imx8m PLL is not able to output a exactly multiple of 44.1 kHz MCLK
input to audio codecs, starting from 24.576 MHz module clock.
If mclk-fs is 256, optimal clock is 11289600,
obtained clock is 11234743. Frequency erro is -0,49 %.
Instead, it is able to generate a exact multiple of 48 kHz MCLK.

Related-to: ELB-4481
Fixes: 9e5c0e3043a7 ("alsa-state: set up asound configs dynamically for verdin-imx8 boards")
Signed-off-by: Emanuele Ghidoli <emanuele.ghidoli@toradex.com>