machine: verdin-imx8mp: include nxp89xx firmware
Currently the nxp89xx firmware is being included at the image level
(i.e. when Toradex BSP users build a Toradex reference image).

Because of this, custom images that are not based on the Toradex
reference image don't get the nxp89xx firmware, and WiFi doesn't work.

Fix that by including the firmware as a RRECOMMENDS package for the
machine.

Obs.: This change affects only images that are based on
packagegroup-base, which does not include for example the
core-image-minimal.

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>