machines: fix fitimage ramdisk load address
Booting with a FIT image (that includes a ramdisk image) is failing on
several machines with the following error message:

Error: ramdisk overwritten
Ramdisk image is corrupt or invalid

This is because the boot script uses ramdisk_addr_r to load the FIT
image, and this is overlapping with the default fitimage ramdisk load
address.

Fix by adjusting the fitimage ramdisk load address:

UBOOT_RD_LOADADDRESS = ramdisk_addr_r + 32M

Build and boot time tested on all impacted machines.

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>