machines: arch arm: follow change to vendor subdir for dtbs
With moving to kernel 6.5 the kernel commit 724ba67515320 ("ARM: dts:
Move .dts files to vendor sub-directories") introduced a regression
with our 32 bit machines.
We have to add the relative path to the dts/dtb files in our list
of device trees to be built per machine if the distro/kernel flavour
is building a 6.5 and later kernel.

Reassign the full list in the <MACHINE.inc> file to KERNEL_DEVICETREE,
so that we can conditionally add the relative path for a 6.5 kernel.

Related-to: ELB-5263
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>