machines: arch arm64: use var to set relative path to dtb
Follow the pattern used in commit ce9a5ef ("machines: arch arm: follow
change to vendor subdir for dtbs")
Set the variable KERNEL_DTB_PREFIX with the relative path to the device
tree source files and then use it. This variable can then be used
in the creation of the u-boot-distro-boot script.

Related-to: ELB-5263
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>