linux-toradex: verdin-imx8mm/mp: disable dsi and lvds upstream drivers
In the BSP7 kernel (toradex_6.6-2.0.x-imx), both upstream and downstream
drivers are included for some devices (two different drivers with the
same DT compatible). With the centralization of the BSP Kconfig, both
upstream and downstream configurations are now enabled simultaneously,
which can cause conflicts between these drivers. For this patch, disable
the upstream DSI Host and the native LVDS controller drivers used with the
iMX8MM and iMX8MP.

Related-to: ELB-5849
Signed-off-by: João Paulo Gonçalves <joao.goncalves@toradex.com>