u-boot-toradex: imx: inherit uuu_bootloader_tag
meta-freescale in commit 161f1b3e69a3 ("boot: Fix UUU tagging, extend
to fslc") requires the boot firmware to be present also in
*.tagged file.

Inheriting uuu_bootloader_tag in the U-Boot recipe creates that
additional file if U-Boot's binman is used.
As the downstream U-Boot is only used for Apalis iMX8 and Colibri iMX8X
where we currently cannot use binman the change prepares for when we
might start to use binman.

The *.tagged file has an appended magic string and is used for wic
images. It allows uuu to extract the boot firmware from the wic image.

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>