imx-oei: use toradex fork
The Optional Execution Image is a plugins to the iMX95 BootROM.
It is mainly used to load the ddr controller firmware, configure
and train the memory.

The project was forked to add the Toradex SMARC iMX95 SoM support.
Override the recipe to use it.

Related-to: ELB-6004
Signed-off-by: Emanuele Ghidoli <emanuele.ghidoli@toradex.com>
(cherry picked from commit d9a1179e4c54e738e3fae83c06217bf01571ead4)