imx-system-manager: use toradex fork
The System Manager (SM) is an application that runs on a Cortex-M processor.
The Cortex-M is the boot core, runs the boot ROM which loads the SM
(and other boot code), and then branches to the SM.
The SM then configures some aspects of the hardware such as isolation
mechanisms and then starts other cores.
After starting these cores, it enters a service mode where it provides
access to clocking, power, sensor, and pin control via a client RPC API
based on ARM's System Control and Management Interface (SCMI).
To facilitate isolation between cores, the SM partitions the SoC into
logical machines (LM) which have statically configurable access rights
to both hardware and RPC API calls.

The project was forked to add the Toradex SMARC iMX95 SoM support.
Override the recipe to use it.

Related-to: ELB-6013
Signed-off-by: Emanuele Ghidoli <emanuele.ghidoli@toradex.com>