linux-firmware: add bbappend recipe to bump version to 20250311
Get the latest linux-firmware tag, 20250311. This is a workaround to
solve an unexpected BLE connection timeout issue experienced with the
older firmware, 20240909, the current version of oe-core scarthgap
branch.

The patch to bump the oe-core to 20250311 was already sent to oe-core
mailing list, therefore this recipe should be removed once the patch
is merged upstream.

Related-to: ELB-6037
Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>