linux-firmware: fix do_install
The bump to version 20250311 also brought an update to the
copy-firmware.sh script that is used as part of the firmware
installation process. It now supports a "-j" flag for parallel
installation, but only if the parallel package is installed. This is not
the case for us, leading to the following output when do_install is ran:

ERROR: the GNU parallel command is required to use -j

The upstream linux-firmware_20250311.bb deals with this by setting
PARALLEL_MAKE="" so the "-j" flag is not passed. Add this workaround to
our bbappend as well.

Fixes: 16e3fa520440 ("linux-firmware: add bbappend recipe to bump version to 20250311")
Related-to: ELB-6037
Signed-off-by: Ernest Van Hoecke <ernest.vanhoecke@toradex.com>