machine: toradex-smarc-imx8mp: remove dsi-to-hdmi_overlay from default overlays
Weston display cloning is not enabled, and the current configuration
points to the first HDMI port, which is the DSI-to-HDMI interface.
As a result, qt-tezi starts on this unused port instead of the native HDMI.

Remove the DSI-to-HDMI overlay from the default overlays list to ensure
that only the native HDMI is used.

Related-to: TEI-1246 TEI-1228
Fixes: fa355bfdb3e1 ("machine: toradex-smarc-imx8mp: Add default overlays")
Signed-off-by: Emanuele Ghidoli <emanuele.ghidoli@toradex.com>
(cherry picked from commit 60c214e809891b88bdb37ee1d42213c85c6a62cb)