verdin-imx95: Use LPDDR_FW_VERSION
The version of the LPDDR FW now differs from the default version
specified by the mkimage Makefile. Therefore, we have to pass the
correct version of the FW we install to mkimage to be able to build.

The LPDDR_FW_VERSION variable is used for this by imx-boot recipe in
meta-freescale.

Related-to: ELB-6342
Fixes: 7f65ae077797 ("verdin-imx95: add machine")
Signed-off-by: Ernest Van Hoecke <ernest.vanhoecke@toradex.com>
(cherry picked from commit 6980a5e91e17ac4e24c47ceb5fdaddea011f089b)