toradex-smarc-imx95: Use LPDDR_FW_VERSION
meta-freescale updated imx-boot to NXP BSP 6.12.3_1.0.0.
The version of the LPDDR FW now differs from the default version
specified by the imx-mkimage Makefile. Therefore, we have to pass the
correct version of the FW we install to mkimage to be able to build.

The LPDDR_FW_VERSION variable is used for this by imx-boot recipe in
meta-freescale.

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>