toradex-smarc-imx95: add toradex-smarc-imx95-a1
Soon, the iMX95 Rev A0 and A1 will be superseded by Rev B0. The B0
revision will become the new default in the near future and requires
some changes.

To support the A0 and A1 silicon, create a new machine for those boards.

For now, toradex-smarc-imx95 still targets A0/A1 silicon with this
commit and both machines are identical, but that will change.

Signed-off-by: Ernest Van Hoecke <ernest.vanhoecke@toradex.com>