machine: verdin-imx95: fix ramdisk load address for FIT image
Move the ramdisk load address 32MB ahead to prevent a memory overlap
between the kernel FIT image and the ramdisk, as previously done with
the toradex-smarc-imx95; with this, both machines keep the same memory
layout. This fix is relevant when a FIT image is in use such as with
Secure Boot images.

Related-to: TOR-3932
Fixes: 7f65ae077797 ("verdin-imx95: add machine")
Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>