kernel-module-isp-vvcam: drop bbappend
From kernel 6.6 to 6.12 was an API change. The isp-vvcam module from
the 6.12 BSP no longer compiles against a 6.6 kernel which was fixed
with a patch.
commit bea597060780 ("kernel-module-isp-vvcam: fix build against nxp
6.6 kernel")

With the bump to a NXP 6.12 based kernel the patch is no longer needed.
Drop the bbappend which applies the patch.

Related-to: ELB-6572
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>
(cherry picked from commit aade6e91ef83ecaef0196ece5feb95cecb51e3cf)