verdin-imx8mp: remove hdmi overlay
With the move to downstream kernel 6.12 the native HDMI output is now
enabled in the default device tree and the overlay to enable it was
removed from the device tree overlays.

Thus remove it from the overlays enabled by default which makes the
upstream and downstream list the same.

Related-to: ELB-6572
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>
(cherry picked from commit 710196d3a09070ffbb2dd216b4a93ce8bf54c32e)